refactor(scorecard): move jira integration constants from scorecard-common to the plugin folder
