+++ workspaces/scorecard/plugins/scorecard-backend-module-jira/src/clients/JiraCloudClient.test.ts
@@ -17,7 +17,7 @@
 import type { Config } from '@backstage/config';
 import type { Entity } from '@backstage/catalog-model';
 import { JiraCloudClient } from './JiraCloudClient';
-import { ANNOTATION_JIRA_PROJECT_KEY } from '@red-hat-developer-hub/backstage-plugin-scorecard-common';
+import { ANNOTATION_JIRA_PROJECT_KEY } from '../constants';
 
 global.fetch = jest.fn();
