+++ workspaces/scorecard/plugins/scorecard-backend-module-jira/src/constants/jiraOpenIssues.ts
