+++ workspaces/scorecard/plugins/scorecard-common/src/index.ts
@@ -16,4 +16,3 @@
 
 export * from './types';
 export * from './permissions';
-export * from './constants';
